pyproject.toml
[project]
name = "fastapi-opinionated-starter"
version = "0.1.0"
description = ""
authors = [
    {name = "Azzarnuji",email = "azzarnuzy93@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi-opinionated-core @ file:///home/zarnuji/project/my-framework/fastapi-opinionated-core"
]


[tool.poetry]
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
fastapi-opinionated-core = {develop = true}

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
